+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `RSpec/NoExpectationExample`. ([@r7kamura][])
 * Add some expectation methods to default configuration. ([@r7kamura][])
 * Fix a false positive for `RSpec/Capybara/SpecificMatcher`. ([@ydah][])
+* Fix a false negative for `RSpec/Capybara/SpecificMatcher` for `have_field`. ([@ydah][])
 
 ## 2.12.1 (2022-07-03)